% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{cond_update_internal_data}
\alias{cond_update_internal_data}
\title{Conditionally update internal data}
\usage{
cond_update_internal_data(
  force = FALSE,
  dl_oecd = TRUE,
  dl_imf = TRUE,
  dl_imfgdpd = TRUE
)
}
\arguments{
\item{force}{Logical. If TRUE, forces the update regardless of data age. Default is FALSE.}
}
\value{
Path to the temporary data directory
}
\description{
Checks the age of the temporary data directory and updates it if older than 7 days,
or if `force` is set to TRUE.
}
