% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_helpers.R
\name{country_cleaner}
\alias{country_cleaner}
\title{Clean country names to match between IMF/ OECD ISO style country names and user input.}
\usage{
country_cleaner(data)
}
\arguments{
\item{data}{A `character`}
}
\description{
`country_cleaner` recode country names form english to ISO3c format
}
\details{
This is a helper function for clean up country names.
}
