% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costdeflate.R
\name{country_code_list}
\alias{country_code_list}
\title{Country code list}
\usage{
country_code_list()
}
\value{
Returns a \tibble{tibble}.
}
\description{
`country_code_list` returns a returns a list of available countries in the WEO data set.
}
\details{
This is a helper function for checking allowable country mappings in the `countrycode` package between English and iso3c standards.
}
