% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costdeflate.R
\name{deflator_country_list}
\alias{deflator_country_list}
\title{Deflator country list}
\usage{
deflator_country_list()
}
\value{
Returns a \code{chr}.
}
\description{
`deflator_country_list` returns a returns a list of available countries in the WEO data set.
}
\details{
This is a helper function for checking which countries exist in the reference data
}
