% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costdeflate.R
\name{delfator_country_year_combs}
\alias{delfator_country_year_combs}
\title{Deflate country/ year combo list}
\usage{
delfator_country_year_combs()
}
\value{
Returns a \code{chr}.
}
\description{
`delfator_country_year_combs` returns a returns a list of available countriy/ year combinations in the WEO data set.
}
\details{
This is a helper function for checking which countr/ year combinations  exist in the reference data
}
