% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_access.R
\name{get_imf_ppp}
\alias{get_imf_ppp}
\title{Get IMF GDP_D values}
\description{
Get Implied PPP conversion rate (PPPEX) values from IMF data sets
}
\details{
Internal function to be used by developers to update internal (i.e.: redundant IMF PPPEX values) and also by the user when loading the package to ensure up to date values are used in the `deflate()` function. see <a></a> for more information
}
