% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_access.R
\name{get_oecd_ppp}
\alias{get_oecd_ppp}
\title{Get OECD PPP values}
\description{
Get Purchasing power parity (PPP) values from OECD data sets
}
\details{
Internal function to be used by developers to update internal (i.e.: redundant OECD PPP values) and also by the user when loading the package to ensure up to date values are used in the `deflate()` function. see <a></a> for more information
}
