% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{imf_gdpd}
\alias{imf_gdpd}
\title{IMF GDP Deflator Values}
\format{
A tibble with 9,292 rows and 3 variables:
\describe{
  \item{COUNTRY}{Country name - WEO format}
  \item{YEAR}{Year}
  \item{NGDP_D}{Gross Domestic Product Price Deflator index}
}
}
\source{
<https://www.imf.org/-/media/Files/Publications/WEO/WEO-Database/2023/WEOOct2023all.ashx>
}
\usage{
imf_gdpd
}
\description{
A dataset containing Gross Domestic Product Price Deflator index values for various countries and years.
}
\keyword{datasets}
