% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{imf_ppp}
\alias{imf_ppp}
\title{IMF Purchasing Power Parity Values}
\format{
A tibble with 9,009 rows and 3 variables:
\describe{
  \item{COUNTRY}{Country name - WEO format}
  \item{YEAR}{Year}
  \item{PPPEX}{Purchasing power parity value}
}
}
\source{
<https://www.imf.org/-/media/Files/Publications/WEO/WEO-Database/2023/WEOOct2023all.ashx>
}
\usage{
imf_ppp
}
\description{
A dataset containing purchasing power parity values for various countries and years from the IMF.
}
\keyword{datasets}
