% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oecd_ppp}
\alias{oecd_ppp}
\title{OECD Purchasing Power Parity Values}
\format{
A tibble with 2,079 rows and 3 variables:
\describe{
  \item{COUNTRY}{Country name - WEO format}
  \item{YEAR}{Year}
  \item{PPP}{Purchasing power parity value}
}
}
\source{
<https://www.oecd.org/en/data/indicators/purchasing-power-parities-ppp.html>
}
\usage{
oecd_ppp
}
\description{
A dataset containing purchasing power parity values for various countries and years from the OECD
}
\keyword{datasets}
