% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{safe_fetch}
\alias{safe_fetch}
\title{Prepare internal data in temporary directory}
\usage{
safe_fetch(fetch_fn, fallback_name, filename, dir)
}
\arguments{
\item{fetch_fn}{function to use to fetch data from api.}

\item{fallback_name}{name of internal dataset to use as fallback.}

\item{filename}{name of file in temp folder.}

\item{dir}{character string contain temp directory path.}
}
\value{
Path to the temporary data directory
}
\description{
Downloads OECD and IMF datasets and stores them in a temporary directory.
Falls back to internal data if download fails.
}
