% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data}
\format{
A tibble with 23 rows and 5 variables:
\describe{
  \item{Reference}{Reference of study}
  \item{Field}{Name of field/cost}
  \item{country_orig}{Country name for cost}
  \item{year_orig}{Year of cost}
  \item{cost_orig}{Cost}
}
}
\source{
<https://doi.org/10.1016/j.chest.2023.06.040>
}
\usage{
test_data
}
\description{
A dataset containing test data for reference and testing.
}
\keyword{datasets}
