% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{update_internal_data}
\alias{update_internal_data}
\title{Prepare internal data in temporary directory}
\usage{
update_internal_data(dl_oecd = TRUE, dl_imf = TRUE, dl_imfgdpd = TRUE)
}
\value{
Path to the temporary data directory
}
\description{
Downloads OECD and IMF datasets and stores them in a temporary directory.
Falls back to internal data if download fails.
}
