% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{cond_update_internal_data}
\alias{cond_update_internal_data}
\title{Conditionally update internal data}
\usage{
cond_update_internal_data(
  dir,
  force = FALSE,
  dl_oecdppp = TRUE,
  dl_imfppp = TRUE,
  dl_imfgdpd = TRUE
)
}
\arguments{
\item{dir}{Character. A path to store data. If missing, will use a temp dir.}

\item{force}{Logical. If TRUE, forces the update regardless of data age.}

\item{dl_oecdppp}{Logical. Whether to check/update OECD PPP data.}

\item{dl_imfppp}{Logical. Whether to check/update IMF PPP data.}

\item{dl_imfgdpd}{Logical. Whether to check/update IMF GDP deflator data.}
}
\value{
Path to the temporary data directory
}
\description{
Checks the age of each dataset in the temporary directory and updates only those
that are missing or older than 7 days, or if `force` is set to TRUE.
}
