% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costdeflate.R
\name{deflate}
\alias{deflate}
\title{Deflate currency values}
\usage{
deflate(
  input_data = .x,
  cost_base,
  year_base,
  country_base,
  year_target,
  country_target,
  cost_target = "cost_target",
  rename_countries = FALSE,
  use_live_data = TRUE
)
}
\arguments{
\item{input_data}{A `data.frame`.}

\item{cost_base}{A numeric or column name representing the base cost.}

\item{year_base}{A numeric or column name representing the base year.}

\item{country_base}{A `character` the name of a country OR the column name representing the base country.}

\item{year_target}{A numeric or column name representing the target year.}

\item{country_target}{A character or column name representing the target country.}

\item{cost_target}{String, what should the name be for the new column with deflated cost values.}

\item{rename_countries}{A Logical
Default to \code{"cost_target"}}
}
\value{
Returns a dataframe with deflated cost values.
}
\description{
`deflate` returns a dataframe with up to 5 additional columns, and main calculates deflated cost values
}
\details{
Deflate currency values.


This is a function for deflating currency, specifying either a base and target country/
year on each row, or applying a pre specified country/ year for each row.
}
\examples{
# Sample dataframe
data <- data.frame(
  country_base = c("USA", "Germany"),
  year_base = c(2010, 2015),
  country_target = c("France", "Italy"),
  year_target = c(2015, 2020),
  cost_base = c(100, 200)
)

# Deflate currency values
deflated_data <- deflate(
  input_data = data,
  cost_base = "cost_base",
  year_base = "year_base",
  country_base = "country_base",
  year_target = "year_target",
  country_target = "country_target"
)

# You can also also specify values in the function for year and country
data <- data.frame(
  cost_base = c(100, 200)
)

# Deflate currency values with pre specified values
deflated_data <- deflate(
  input_data = data,
  cost_base = "cost_base",
  year_base = "2010",
  country_base = "Australia",
  year_target = "2020",
  country_target = "United States"
)

}
