% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costdeflate.R
\name{deflate}
\alias{deflate}
\title{Deflate currency values}
\usage{
deflate(
  input_data,
  cost_base,
  year_base,
  country_base,
  year_target,
  country_target = "USA",
  cost_target = "cost_target",
  pppex_src = "IMF",
  rename_countries = TRUE,
  use_live_data = TRUE,
  force_live_data = FALSE
)
}
\arguments{
\item{input_data}{A `data.frame`.}

\item{cost_base}{A `numeric` or `colname` representing the base cost.}

\item{year_base}{A `numeric` or `colname` representing the base year.}

\item{country_base}{A `character` the name of a country OR the `colname` representing the base country.}

\item{year_target}{A `numeric` or `colname` representing the target year.}

\item{country_target}{A `character` or `colname` representing the target country.}

\item{cost_target}{A `character`, what should the name be for the new column with deflated cost values.}

\item{pppex_src}{A `character`, which dataset should PPP values come from? IMF or OECD?}

\item{rename_countries}{A `logical`, ask the function to clean up country names if not in `iso3c` format.}

\item{use_live_data}{A `logical`. Makes call to `update_internal_data()`, if current IMF or OECD are more than a week old then make an API call to replace them., default = `TRUE`}

\item{force_live_data}{A `logical`. Makes call to `update_internal_data()` regardless of age of data currently stored. Use only if you know IMF WEO or OECD data has been updated since you last ran `deflate`, default = `FALSE`
Default to \code{"cost_target"}}
}
\value{
Returns a dataframe with deflated cost values.
}
\description{
`deflate` returns a dataframe with up to 5 additional columns, and main calculates deflated cost values
}
\details{
Deflate currency values.


This is a function for deflating currency, specifying either a base and target country/
year on each row, or applying a pre specified country/ year for each row.
}
\examples{
# Sample dataframe
data <- data.frame(
  country_base = c("USA", "Germany"),
  year_base = c(2010, 2015),
  country_target = c("France", "Italy"),
  year_target = c(2015, 2020),
  cost_base = c(100, 200)
)

# Deflate currency values
deflated_data <- deflate(
  input_data = data,
  cost_base = "cost_base",
  year_base = "year_base",
  country_base = "country_base",
  year_target = "year_target",
  country_target = "country_target"
)

# You can also also specify values in the function for year and country
data <- data.frame(
  cost_base = c(100, 200)
)

# Deflate currency values with pre specified values
deflated_data <- deflate(
  input_data = data,
  cost_base = "cost_base",
  year_base = "2010",
  country_base = "Australia",
  year_target = "2020",
  country_target = "United States",
  rename_countries = TRUE
)

}
