% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_access.R
\name{get_imf}
\alias{get_imf}
\title{Get IMF data}
\description{
Get helper function for accessing IMF data through SMDX API.
}
\details{
Internal function to be used by developers to update internal (i.e.: redundant IMF PPPEX & GDP deflator values) and also by the user when loading the package to ensure up to date values are used in the `deflate()` function. see <a href = "https://www.imf.org/en/Publications/SPROLLS/world-economic-outlook-databases#sort=%40imfdate%20descending">IMF WEO website</a> for more information.
}
