% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_access.R
\name{get_imf_gdpd}
\alias{get_imf_gdpd}
\title{Get IMF GDP_D values}
\description{
Get gross domestic product, deflator (GDP_D) values from IMF data sets
}
\details{
Internal function to be used by developers to update internal (i.e.: redundant IMF GDP values) and also by the user when loading the package to ensure up to date values are used in the `deflate()` function. see <a href = "https://www.imf.org/en/Publications/SPROLLS/world-economic-outlook-databases#sort=%40imfdate%20descending">IMF WEO website</a> for more information.
}
