% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{safe_fetch}
\alias{safe_fetch}
\title{Prepare internal data in temporary directory}
\usage{
safe_fetch(fetch_fn, fallback_name, filename)
}
\value{
Path to the temporary data directory
}
\description{
Downloads OECD and IMF datasets and stores them in a temporary directory.
Falls back to internal data if download fails.
}
